refactor: remove nextjs folder structure
