+++ platform/src/app/assignment/components/section-container.tsx
