+++ platform/src/app/assignment/components/section-trigger.tsx
