+++ platform/src/app/assignment/components/transcriber-section.tsx
