+++ platform/src/app/assignment/components/validator-section.tsx
