+++ platform/src/app/courses/components/course-card-skeleton.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Card, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { Skeleton } from "@/components/ui/skeleton";
