+++ platform/src/app/courses/components/course-card.tsx
