+++ platform/src/app/courses/components/course-form-dialog.tsx
