--- platform/src/app/demo/page.tsx
@@ -1,12 +1,0 @@
-"use client";
-import { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-
-export default function Page() {
-  const navigate = useNavigate();
-  useEffect(() => {
-    navigate("/demo/courses", { replace: true });
-  }, [navigate]);
-  return <></>;
-}
-
