--- platform/src/app/layout.tsx
@@ -1,37 +1,0 @@
-import type { Metadata } from "next";
-import { Geist_Mono } from "next/font/google";
-import { hostGrotesk, remark } from "@/lib/fonts";
-import "./globals.css";
-import Header from "@/components/header";
-import {Providers} from "./providers";
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Fair Platform",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-                                     children,
-                                   }: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-    <body
-      className={`${geistMono.variable} ${hostGrotesk.variable} ${remark.variable} font-sans antialiased h-screen`}
-    >
-    <Providers>
-      <Header />
-      <div className={"pt-16 h-full"}>
-        {children}
-      </div>
-    </Providers>
-    </body>
-    </html>
-  );
-}
